requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
prompt_toolkit==3.0.50
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
urllib3==2.3.0
wcwidth==0.2.13
